requirements.txt
easydict==1.9
PyYAML==5.3.1
tqdm==4.45.0
